Fix a potential crash when changing to fish form with huggies attached.
This patch corrects a bug which can cause the game to crash when Naija
changes to fish form while huggies are attached, because the script
tries to use an already-freed bone pointer.  Another symptom is that the
huggies suddenly disappear when Naija changes form (due to being warped
to invalid coordinates), rather than flying off as usual.

This patch is applied against the new-style script from the lua-changes
branch, but the bug appears to be present in the original script as well.