Add a readFile() function to read a file in one call.
This changeset is a backport of PSP import changesets #9/40
(r33:e15ae10a7c16), which adds a global readFile() utility function
allowing the entire contents of a file to be loaded into memory with
one call, and #12/40 (r36:3ad4a420d7be), which uses readFile() to load
the TrueType font into a memory buffer and reuse that buffer for all
three instances of the font.  The regression fix in r102 (4e1bb91a782a)
is also included.