Implement alternative visited-area display logic.
This changeset is a backport of the texture-write display method used on
the PSP for visited areas on the world map (changesets r89:c63a7bd8973d
and r277:69cdcd5c6e61), without the PSP-specific code.  The new code is
not currently used on other platforms, but it is available if needed.