Backport obstruction grid rendering optimizations from the PSP branch.
This changeset is a backport of the grid rendering optimizations
committed on the PSP branch (r176:619b05f42999, r177:e4b0246ccd9f, and
r184:65ec99dc1ee8).  The primary optimization is inverting the render
loop order to match the memory layout of the obstruction grid, to avoid
cache thrashing.