Fix a pointer type mismatch in an Energy Temple orb/door script.
This patch fixes the improper passing of a Node object to
entity_getNearestEntity() in scripts/maps/node_energytemple01door.lua,
which had the potential to cause a floating-point exception depending on
the contents of memory near the Node object.  This would either crash
the program or result in the orb holder at the lower-right exit of the
EnergyTemple01 area missing its orb despite the door being open.

This bug was exposed by the optimization of the "closest" search in
entity_getNearestEntity() on the use-huge-valf branch; the original
code, which used -1 as a "nothing found yet" flag, would (if the program
didn't crash) accept the first entity despite finding a NaN distance,
resulting in the correct behavior of having an orb in the holder.