Add an up-to-date check for the PSP data files.
This patch adds logic to store a data version number in "version.txt" in
the PSP data directory, and use that file to check whether the data
needs to be regenerated.  The data version number is set to "2" in this
patch; since previous versions did not store a version number, this will
force data generated by previous versions to be regenerated (thus fixing
textures corrupted by the bug resolved in r356:f4a46b5acaa0).