Build an icon for the game when using the Aquaria PSP Builder.
This patch updates the Aquaria PSP Builder to generate a PSP icon file
(ICON0.PNG) for the game from the title logo included with the data from
the commercial release.

Since the icon can now be built along with the rest of the data, the
ICON0_* variables in the PSP Makefile are now disabled by default, and
the EBOOT.PBP generated by the "make" process will not contain an icon
unless one is placed in the PSP directory beforehand (as "ICON0.PNG").