Allow a custom string to be appended to the title screen version number.
This patch adds a CMake variable AQUARIA_CUSTOM_BUILD_ID, allowing a
custom string to be appended to the version number displayed on the
Aquaria title screen.  The string is appended with no intervening space
or other separator, so any needed separator should be included at the
beginning of the string, such as "-custom" or " (custom build)" (which
would display as "1.1.3-custom" or "1.1.3 (custom build)" respectively).

The default value of the variable is the empty string, so behavior will
be unchanged from current builds unless -DAQUARIA_CUSTOM_BUILD_ID="..."
is given on the "cmake" command line or configured in the CMake GUI.