Use angle rather than position for selecting notes with the joystick.
This patch adjusts the joystick input logic for the singing interface
so that the input angle, rather than the resultant position on the note
circle, is used to select the note to sing.  This provides a little more
leeway on detecting the last of a series of consecutive notes, so the
player doesn't have to worry about getting the joystick to exactly the
right angle.