Remove four script functions determined to be no-ops.
This patch removes the following four script functions, which were
determined through code analysis to be no-ops:

- entity_getBehaviorType
- entity_setBehaviorType
- entity_setTouchPush
- setGameOver

Of these, the only one which was in fact called is entity_setTouchPush()
(by scripts/entities/energyboss.lua); that call has been commented out
in this patch.  entity_setBehaviorType() was referenced by the "activate"
callback in blaster.lua and bigblaster.lua, but since those entities can
never be activated, the code would never be reached.  (The callbacks
themselves have been removed in this patch for clarity.)