Various PSP README updates.
This patch updates a few out-of-date notes in the PSP README file,
among other miscellaneous tweaks.  An "IMPORTANT" note has also been
added to the section on the Aquaria PSP Builder tool, reminding users
not to mix executable and data files.  (Since the scripts are stored in
the package file, they must match the script interface in the program,
or some scripts may fail, potentially leaving the game stuck.)