Remove a duplicate definition of STATE_TRANSFORM in the script constant table.
This patch removes the definition of the script constant STATE_TRANSFORM
as the value of Entity::STATE_TRANSFORM, since there is a subsequent
definition with a different value (314).

Note that Entity::STATE_TRANSFORM is not actually used in the C++ code;
there are references to it in Aquaria/Avatar.cpp, but Naija can never be
set to that state, so the associated code is all dead.