Fix incorrect marking of "visited" areas on the world map.
This patch fixes a cosmetic bug which caused the top edge of each area
to be inappropriately displayed as "visited" while moving through the
area (most noticeable in the Veil areas as a white fringe along the top
of the area in the world map).  This is only a bug in the display of the
current area, and does not affect save files; the appearance is restored
to normal after moving to a different area or selecting a different area
on the world map.