Mark "visited" regions on the world map based on what's visible on the screen.
This patch tweaks the logic for marking "visited" regions of the world
map so that the central 80% (by width and height) of the region visible
on the game screen at any given time is marked visited, rather than
marking only the 9 map units of the current area centered on Naija's
position.  Typically, this will result in a slightly wider swath of the
map being marked visited as Naija moves through the area.

This patch is based on the mapvis-bitmap branch, and will not apply
directly to the original sources.