Prevent Li and the pet blaster from hitting seahorses and other non-targets.
This patch alters the collision logic for the energy shots fired by Li
and the pet blaster so that they do not collide with entities which are
explicitly set to ignore those attacks.  For example, this allows Li to
shoot through seahorses and hit enemies on the far side (just as Naija's
energy attack does), rather than hitting and potentially killing the
seahorse.

WARNING:  While this patch resolves what could be considered a bug with
respect to seahorses and other such neutral entities, the effect of the
change is game-wide.  As a result, this patch may introduce regressions
preventing Li or the pet blaster from hitting targets they should
normally be able to hit.