Avoid the orange/purple gem icons on the world map interface overlapping the minimap.
This patch adjusts the positions of the orange and purple pyramid "gem"
icons on the world map interface (used for placing user markers, located
immediately above the minimap) to take into account the displayed size
of the minimap.  This prevents the icons from overlapping the minimap if
the minimap size is set larger than the default for a specific platform.