Fix "dancing dots" on world map beacons at high zoom.
This patch corrects a minor graphical bug in the flickering dots used to
indicate a "beacon" on the world map.  The code for bounding the dots'
positions to the edges of the screen did not take into account the scale
of the world map itself, so when zoomed in to a high scale factor, dots
whose actual position was offscreen would overshoot the edges of the
screen and appear to "flicker" as they bounced back and forth.  (At low
frame rates, a similar issue could be observed when zoomed out, as the
dots would geometrically approach the edge of the screen.)

This was reported as bug #4542 on bugzilla.icculus.org.