Fix a crash on joystick movement immediately after loading.
This patch fixes a regression introduced in r111 (86ae2bc43cea) which
caused the GUI element for the selected save slot in the load menu to
continue accepting joystick input even after the game was loaded.  This
resulted in attempts to apply input focus to destroyed GUI elements,
potentially causing a crash if the player tried to start moving with a
joystick immediately after loading completed.

Note that this patch requires r112:6b5d95f6d124 in order to apply
properly.