Prevent the death sequence from starting while the game is paused.
This patch prevents the Naija death sequence from starting while the
game is paused.  Previously, if the in-game menu, help screen, or world
map were opened at the same instant Naija died, the game would be left
in an inconsistent state after reload, leaving Naija invisible or even
crashing the program.

This was reported on the Bit Blot forums, at (among other places):
http://www.bit-blot.com/forum/index.php?topic=1976.0