Prevent menus from opening while changing areas.
This patch prevents the in-game menu, world map, and help screen from
being opened while a change of game state is taking place (such as when
warping from one area to another).  As with the bug fixed in r440
(35b08766e7b9), opening one of these interfaces at the same instant
Naija exited an area would leave the game in an inconsistent state,
potentially causing a crash.