Prevent shock jellyfish from zooming around the map after being pushed.
This patch fixes a bug which caused shock jellyfish (the orange-yellow
jellyfish that occasionally shock the area around them) to start moving
at very high speeds.  This occurred when Naija boosted into a jelly
immediately before the jelly began preparing a shock attack, because
the shock attack overrode the return to STATE_IDLE which would normally
reset the jelly's maximum speed.

This was reported on the Bit Blot forums:
http://www.bit-blot.com/forum/index.php?topic=1709.0