Prevent the fourth form of the last boss from spawning multiple times.
This patch fixes a bug which could cause the fourth form of the Creator
to spawn multiple times.  This happened if the third form was hit by
multiple energy shots at once when down to its last hit; the script
would spawn one copy of the fourth form for each shot that registered.

This was reported on the Bit Blot forums:
http://www.bit-blot.com/forum/index.php?topic=2036.0