Prevent jellyfish from getting stuck in mid-air.
This patch fixes the entity movement logic so that entities which are
normally not allowed to leave the water can nonetheless continue moving
if they are somehow forced out of the water.  This can happen if the
entity is pushed (as happens to jellyfish when Naija runs into them at
bursting speed), since pushing overrides the can-leave-water check.

This patch also fixes the various jellyfish scripts so that jellyfish
pushed out of the water will properly fall back in.  Of the five
jellyfish varieties, the Nauplius and the "zapping" jellyfish (found
in the Body area) don't currently appear in any area with out-of-water
regions, but the scripts are updated nonetheless for potential use in
mods.

Note that if the program fix is applied but the scripts are not updated,
jellyfish knocked out of the water will continue to "swim" as though
they were still underwater -- which itself is a fairly amusing image...