Update the recipe preview graphic on cooking a previously unknown recipe.
When first cooking a previously unknown recipe, the recipe preview
graphic (on the right side of the menu) was not properly updated to
show the newly learned recipe if ingredients remained in the cooking
slots, instead remaining as a "?".  This patch ensures that the recipe
preview graphic is properly updated in such cases.