Fix a bug causing buggy scripts to be improperly cached.
This patch corrects a logic bug which causes scripts which fail during
initial loading to remain cached in memory, so that the failing version
is used even if the script file is subsequently updated.  (The same
problem would occur if the system ran out of memory trying to create a
Lua thread for the script, though in that case the game itself would
likely crash in short order.)

This fix only affects people who edit scripts while the program is
running, such as mod developers.