Expand the world map UI graphic to allow for slightly larger text.
This patch expands the UI graphic shown at the top of the world map to
allow for slightly larger (and thus more readable) area name labels,
while also avoiding a couple of cases in which long area names could
slightly overlap the graphic itself.

To make room for the expanded UI graphic, the help button (which wasn't
selectable anyway) is removed by this patch.