Resize the world map UI graphic at data file generation time.
This patch modifies the Aquaria PSP Builder to resize the world map UI
graphic to native resolution when the PSP data files are generated,
improving the readability of the area name labels ("Hover", "In", and
"Selected").  The rendering logic in Aquaria/WorldMapRender.cpp is also
updated to match.

Using an "aquaria.dat" built by a previous version of the builder tool
with this version of the engine will result in the world map UI graphic
appearing too large (overflowing the screen).