Fix dark areas shifting position on non-4:3 displays.
This patch corrects a logic bug which caused "darkness" effects on maps
to shift position as the map scrolled.  On a widescreen display, for
example, the darkened areas would move leftward relative to the rest of
the map as Naija swam to the right; this is particularly noticeable near
the save point before the Energy Temple boss.