Fix a harmless logic glitch in the song menu joystick tweak.
This patch fixes a logic glitch introduced in r254 (1dd8c3be1a97), which
changed the song menu handling so that the secondary button could also
be used to trigger the song description (for ease of use with joysticks).
The glitch has essentially no actual impact; at worst, it could cause a
very short button press to be missed.