Renumber PSP save directories to start from 0 instead of 1.
This patch changes the numbering of the save file directories used on
the PSP so that the directory names count from 0 instead of 1; for
example, save #1 is now "GAME00000Aquaria_000" instead of "..._001".
This will naturally cause all existing save files to shift down by one
slot in the save screen, so that a game previously saved in slot #1 will
now appear in slot #2, and a game previously saved in slot #32 will no
longer be loadable.

Due to the DRM used in PSP save files, it is unfortunately not possible
to rename the directories from a PC; be sure in particular that if you
have a game saved in slot #32, you load and save it in a different slot
_before_ applying this patch.

This change was required due to interface changes in the new Aquaria
engine (currently under development), so I've decided to make the same
change here on the "better sooner than later" principle.