Tweak the progress display during the final build stage of the PSP builder.
This patch adjusts the progress calculation during the aquaria.dat build
stage of the AquariaPSPBuilder tool, so that the sizes of the data files
being copied are taken into account.  This helps reduce the slowdown of
the progress bar seen when the (comparatively large) music files are
copied, though it also results in the progress bar advancing abnormally
quickly if the files' data is already cached.