Allow system libraries to be used in place of the included versions.
This patch updates the CMake control file CMakeLists.txt to detect the
presence of the FreeType, Lua, and Ogg/Vorbis libraries on the system,
and (if found) use those libraries in place of the copies included in
the ExternalLibs directory of the Aquaria source tree.  To force the use
of the included libraries, the following CMake variables can be set:

    AQUARIA_INTERNAL_FREETYPE
    AQUARIA_INTERNAL_LUA
    AQUARIA_INTERNAL_OGGVORBIS

This patch is based on the move-external-libs branch, though a similar
patch could be applied to the original icculus.org source tree.