Move more unused scripts to the "_unused" directories.
This changeset expands on r227:355198e6ac29 by checking more extensively
which entity scripts are in fact used, either in a createEntity() call
or by virtue of being named in one of the map definition files.  Of
these, the "ompo" script in particular (and the supporting "spawnompo"
and "eatompo" node scripts) are widely referenced, but none of the code
is actually executed because the first node in the "TrainingCave" map
which sets the initial flag does not exist; presumably, this is a
feature added during development which was later removed.

In addition to the entity scripts themselves, a few references to the
unused scripts have also been deleted for completeness.