Track elapsed time properly in sprite animation looping.
This patch corrects a bug in sprite animation loop handling which
discarded all elapsed time beyond the end of the current loop when
looping back to the beginning of an animation.  In particular, this
broke the apparently-intended effect of varying the initial positions
of the bird wings in the Bit Blot logo screen.