Ensure consistent behavior when firing shots under joystick control while motionless.
This patch adjusts the logic for firing shots to ensure that unaimed
shots are always fired in the direction Naija is facing when she is
not moving.

Previously, if Naija was not moving and the cursor was located exactly
on Naija's current position -- as would happen when using joystick
control -- shots would behave differently depending on the current form.
Notably, beast form shots would remain motionless at the location where
they were fired.