Implement the "weakened shield" effect for the Shield song.
This patch corrects a logic bug in the handling for the "aura" effect
displayed around Naija while the Shield song is active.  Based on the
code, the aura effect was intended to change to a lower-intensity,
red-tinged "weakened" aura shortly before it expired; however, an
uppercase/lowercase mismatch prevented the effect from ever being used.
With this patch, the effect is displayed properly, and the code is also
slightly improved to prevent a momentary disappearance of the regular
aura that would have occurred at the point of changeover.

While this patch can be viewed as a bugfix, it has been applied on the
gameplay-changes branch because the additional knowledge provided by the
"weakened" effect may change players' behavior.