Tweak the color quantizer used in generating PSP texture files.
This patch adjusts the behavior of the color quantizer used for texture
data generation by the AquariaPSPBuilder tool, fixing some cases in
which transparent portions of textures could be encoded with the wrong
pixel value (either a non-transparent value, or a transparent value with
the wrong color).  Among other things, this corrects the white fringes
previously visible along the edges of the painting in the intro sequence.