Fix a regression causing glitches in rendering segmented entities.
This patch fixes a regression, first introduced in r140 (eec26e072fb6)
on the PSP branch and later merged in r320 (5e02898f53c2) on branch
psp-backport, which caused segmented entities such as the worms in
Mithalas Cathedral to be rendered improperly (such as with spaces
between the segments or with segments facing in wrong directions).
The regression was reported by <andrews05@gmail.com>.

The patch also corrects a missed minor optimization, though the compiler
should normally be able to detect the optimization chance automatically.