Fix a regression in the display of the Shield song effect.
This patch corrects a regression introduced by the addition of the
"weakened shield" effect (in r552:c11a97e8aebc).  The "weakened" effect
was not properly disabled if the player sang the Shield song before the
effect expired, so both the regular and weakened effect would be shown
on top of each other, resulting in a brighter-than-usual aura around
Naija.