Fix a possible entity targeting regression due to a shadowed variable.
This patch corrects a possible regression introduced in r456
(745b77c3735c) when converting entity target management from an array to
a single pointer, caused by failing to catch a local variable that
shadowed the new field name.  The previous code is definitely incorrect,
but I haven't noticed any obvious effects while playing, so it may be a
harmless issue.