Add more keyboard controls to the animation editor.
This patch adds the following keyboard controls to the animation editor:

- PgUp/PgDn: Select the previous/next animation
- Enter: Start the current animation
- Shift+Enter: Stop any running animation

The keyboard controls for changing the sprite scale factor (formerly
PgUp and PgDn) are now Shift+PgUp and Shift+PgDn.

This patch also inverts the positions of the previous/next animation
buttons on screen, putting "prevAnim" on top to match the order of the
PgUp/PgDn keys on the keyboard (as well as the previous/next keyframe
buttons).