Fix an outdated error message.
This patch removes the outdated reference to entityinclude.lua reported
when the watch() function is called with a time parameter of zero.
(The message was presumably included because calling watch(FRAME_TIME)
without first including entityinclude.lua to define FRAME_TIME would
result in calling watch(0), potentially causing an infinite loop if the
script was waiting for something to happen.  Since FRAME_TIME and other
common constants are now automatically defined for all scripts, the
message is no longer relevant.)