Allow display lists to be deactivated in the preferences file.
This patch adds a "displaylists" attribute to the <ScreenMode> tag of
the <Video> block in the preferences file (usersettings.xml), allowing
the use of OpenGL display lists to be selected at runtime.  The default
value of this attribute is 1 (display lists enabled); by setting it to
0, the use of display lists will be inhibited, which may improve
performance with low-end video hardware or software renderers.